+++ src/main/java/lagatrix/gui/views/main/form/UserView.java
@@ -6,12 +6,12 @@ import lagatrix.entities.actions.ActionsEnum;
 import lagatrix.entities.connection.Request;
 import lagatrix.entities.dto.user.User;
 import lagatrix.gui.components.complex.containers.RowContainer;
-import lagatrix.gui.dialog.ErrorDialog;
-import lagatrix.gui.dialog.QuestionDialog;
+import lagatrix.gui.dialog.status.ErrorDialog;
+import lagatrix.gui.dialog.status.QuestionDialog;
 import lagatrix.gui.views.formulary.UserFormularyView;
 import lagatrix.gui.views.main.getters.Getter;
 import lagatrix.gui.views.main.getters.UserGetter;
-import lagatrix.gui.window.FormularyWindow;
+import lagatrix.gui.dialog.FormularyDialog;
 
 /**
  * This view display and manage the users of server.
@@ -158,7 +158,7 @@ public class UserView extends MainView {
     }// </editor-fold>//GEN-END:initComponents
 
     private void addButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_addButtonActionPerformed
-        new FormularyWindow((JFrame) SwingUtilities.getWindowAncestor(this), 
+        new FormularyDialog((JFrame) SwingUtilities.getWindowAncestor(this), 
                 new UserFormularyView(requester)).setVisible(true);
     }//GEN-LAST:event_addButtonActionPerformed
 
@@ -169,7 +169,7 @@ public class UserView extends MainView {
             view = new UserFormularyView(requester, (User) getRowContainer().
                             getSelectedRow().getEntity());
             
-            new FormularyWindow((JFrame) SwingUtilities.getWindowAncestor(this), 
+            new FormularyDialog((JFrame) SwingUtilities.getWindowAncestor(this), 
                     view).setVisible(true);
             
             rowContainer.getSelectedRow().setEntity(view.getEntity());